Extract DualPaneMapper
* Extract DualPaneMapper as a standalone component
* MappingWizardClustersStep will implement instances of DualPaneMapper
and TreeView
* MappingWizardClustersStep will manage clusters state
