+++ app/javascript/components/MappingWizardClustersStep/__mocks__/data.js
@@ -1,0 +1,37 @@
+export const source = {
+  name: 'Source Clusters',
+  count: 4,
+  resources: [
+    {
+      id: 1,
+      name: 'VMWareCluster1'
+    },
+    {
+      id: 2,
+      name: 'VMWareCluster2'
+    },
+    {
+      id: 3,
+      name: 'VMWareCluster2'
+    },
+    {
+      id: 4,
+      name: 'VMWareCluster2'
+    }
+  ]
+};
+
+export const target = {
+  name: 'Target Clusters',
+  count: 2,
+  resources: [
+    {
+      id: 5,
+      name: 'RHVCluster1'
+    },
+    {
+      id: 6,
+      name: 'RHVCluster2'
+    }
+  ]
+};
